First, rough effort to abstract/extract EntityCache
The goal is to remove the static link of Entity (dao package) cache to
transaction (sql package). Step one was to rename to ITransactionManager
and ITransaction, then replace Transaction with ITransaction in all API
signatures. The introduce a DaoTransactionManager and DaoTransaction
that extend the behavior of the sql package transaction *primarily by
composition (delegation). These new classes, being entirely within the
dao package can introduce ICache/EntityCache as a delegate so that the
DaoTransaction isolates the idea of caching entirely. This opens up the
possibilty of user-introduced caching for DAOs.



For transaction scopes that include Entity, we just need to switch from
sql's `transaction {}` labmda to the one in dao package. For existing
code bases, this should be the only change necessary.

These changes appear to be ore disruptive than they are due to the
introduction of ITranactionManager and ITransaction into all of the API
signatures.