Relax typing for Transaction#exec (#1123)
The previous code required a `ColumnType` when only a `IColumnType`
was required.  This imposed additional requirements on authors when
writing manual statements.  Relaxing the requirement to `IColumnType`
makes it both easier to write manual statements and allows reuse of
many of Exposed's existing parts, such as `QueryBuilder` which provides
args as a list of `IColumnType`.