Get defaultIsolationLevel before initTransaction
`defaultIsolationLevel` is a `lazy` value and trying to fetch it after `super.doBegin()` in the `doBegin()` function results in a deadlock type of situation, where one connection is essentially and idle and the code it is calling is asking for another one. Since a connection has been opened at the first instance itself, a scenario arises where the creation request for a connection times out due to the calling code waiting for the parent to release its connection, since it needs to acquire it to understand the default isolation level of the database.