fix: EXPOSED-19 Max timestamp in SQLite not working
In the `create` function in `ResultRow.kt`, `columnType` is null because `(field as? Column<*>)` is null. This causes the object read from the DB to be of type String instead of Timestamp, which results in failing to parse the date because it is not in the proper format for `Instant.parse`.

`field` is of type `org.jetbrains.exposed.sql.Max` when we run the test. This is the relationship between the classes (-> means "is a subclass of"):
Max -> Function -> ExpressionWithColumnType -> Expression
Column -> ExpressionWithColumnType -> Expression

We attempt to cast `Max` to `Column` and that fails because there is no relationship between them. If we cast to `ExpressionWithColumnType`, however, it works.