Fix H2 update() with join
H2 MERGE INTO command allows a WHERE clause to be appended to the query.
If a compatibility mode does not support WHERE clause in merge, the driver
throws an unclear syntax error. Documentation and tests shows that the
only H2 mode to support this clause is Oracle mode.

This catches all unsupported cases and throws an exception with a more
informative message.