feat: EXPOSED-32 Support string function CHAR_LENGTH
Rationale for:

- Function Name: Opted to use `CharLength` instead of `Length` to follow
the ANSI SQL standards, even though all dialects support a function with
a variation on the name 'length'. Also, LENGTH() specifically returns the
length of the string measured in bytes in MySQL and MariaDB, whereas all
versions of CHAR_LENGTH() return the expected single-character count.

- Expression Type: The function only accepts string expressions, so an
explicit cast is required if a number needs to be provided to the function.
Of the supported databases, only MySQL, MariaDB, and Sqlite support a
numerical value argument for implicit casting by the database.