fix: EXPOSED-36 LocalDate comparison in SQLite
Date value was being passed to database as java.sql.Date, stored as a
13-digit integer because the object represents milliseconds since
1970-01-01. Date datatype in SQLite passed to date functions as an
integer must be of a specific format, 10 digits, to represent the number
of seconds since 1970.

Conversion to a recognizable format could be achieved by changing the
dialect functions, like year() e.g.:
STRFTIME('%Y', date_column / 1000, 'unixepoch')
But this would break cases where the same date functions are used for
DateTime values.

Date value is instead passed to database as a string, since multiple
formats are recognized for comparison.