fix: EXPOSED-36 LocalDate comparison in SQLite
LocalDate column representation was set as DATE for all databases.

SQLite does not support a date-specific column type and would instead assign the
type as NUMERIC, based on type affinity, with numbers representing seconds since
1970-01-01.

LocalDate column representation was switched to TEXT to match the 3 SQLite
options available for storing date values, as well as the changed format for
passing values to the database (from java.sql.Date to String).