fix: EXPOSED-48 Incorrect statistics aggregate functions
The aggregate functions for statistics were failing on:
- [SQLite] with 'no such function' exception fron DB. Now they fail early with
an UnsupportedByDialectException.
- [SQL Server] because of unrecognized syntax. Now they use the correct function
name.

It was possible to invoke these extension functions on a string-type column, for
example, and compile, which would result in an SQLException thrown by the DB.
The type is now restricted to accept only numeric values.

Add a new test class for statistics aggregate functions.