feat: EXPOSED-66 Extend partial index to SQLServer and SQLite
If creation of a partial index was attempted using a dialect that doesn't support
it, the filter condition was being ignored and a full index was being created
instead. This has been switched so that no index at all is created, which is
more consistent with existing behavior (e.g. if a typed index is attempted in
an unsupporting dialect, the index is also not created). A warning is also logged.

Test names changed to be more descriptive and internal test function renamed
to be consistent with Table property 'indices'.