feat: EXPOSED-66 Extend partial index to SQLServer and SQLite (#1763)
Add SQLServer and SQLite dialects to filterCondition().

If creation of a partial index was attempted using a dialect that doesn't support
it, the filter condition was being ignored and a full index was being created
instead. This has been switched so that no index at all is created, which is
more consistent with existing behavior (e.g. if a typed index is attempted in
an unsupported dialect, the index is also not created). A warning is also logged.

Override createIndexWithType() and dropIndex() where necessary.

Add unit tests.

Remove unused property and function in previous partial index test & change
naming style.