feat: EXPOSED-43 Add support for timestamp with time zone (#1787)
The data type timestamp with time zone is supported by the following databases:

-H2
-Oracle
-SQLite
-SQL Server
-PostgreSQL (converted to UTC when stored, without preserving time zone info)
-MySQL 8.0.19+ (converted to UTC when stored, without preserving time zone info)

It's not supported by the following databases:

-MariaDB

Note: PostgreSQL and MySQL always store the timestamp in UTC, thereby losing the original time zone. To preserve the original time zone, store the time zone information in a separate column.