test: Fix failing datetime tests in MariaDB
The following tests in exposed-java-time and exposed-kotlin-datetime fail when
run using MariaDB:

KotlinTimeTests/'test string LocalDateTime with nanos'()
JavaTimeTests/'test string LocalDateTime with nanos'()

- Fails with the message `Failed on microseconds` because Exposed's assertion
functions for the fractional part are using `roundToMicro()` to compare nanoseconds.
In other databases (DB), including MySQL, the nanoseconds retrieved from the DB
is rounded to the internal precision using RoundingMode.HALF_UP. MariaDB always
rounds down.
- These tests are flaky because they rely on Clock.System.now() (and the java
equivalent) and add a set amount of nanoseconds to a dynamic datetime value.
This means any resulting value on the lower half of a microsecond will be round
down and pass.
- The assertion has been changed to floor the nanoseconds value instead and the
test now takes 2 values with constant nanoseconds to evaluate what happens when
the value is low versus high.