test: Fix failing testSqrtFunction test for SQLite
This test fails locally for me, but not on TeamCity for some reason. According to SQLite documentation, using the SQRT math function on a negative value returns NULL, but there is no mention of an error being thrown, so we need not expect an exception. I tried it on DataGrip and using exec function to execute raw SQL in the test, and both did not throw an error and simply returned null.