fix: EXPOSED-82 Inaccurate UShort column type mapping (#1799)
* fix: EXPOSED-82 Inaccurate UShort column type mapping

Currently, when attempting to insert a UShort value outside of the range [0, 32767],
Exposed truncates the value by calling value.toShort() before sending it to the
DB, causing overflow. The value is stored successfully as a negative number because
all databases (except MySQL and MariaDB) don't support unsigned types natively,
which means Exposed is actually mapping to 2-byte `SMALLINT`, which accepts
the range [-32768, 32767].

Change the default mapping to the next higher-up integer data type INT (technically
a 4-byte storage type) and remove the truncation conversions so that an accurate
value is sent/received to/from the database. To ensure that the intended behavior
cannot be overriden using exec() directly, a check constraint is auto-applied to
the column when registered if the database is not MySQL/MariaDB.

Oracle INT is an alias for NUMBER(38), so this has been overriden to a reduced
type NUMBER(6).