test: Fix failing datetime comparison tests in Oracle (#1813)
* test: Fix failing datetime comparison tests in Oracle

In both exposed-java-time and exposed-kotlin-datetime modules:

MiscTableTest/testSelect01()
MiscTableTest/testSelect02()

Fails with an internal equality comparison issue due to precision.

The datetime() column maps to TIMESTAMP without specifying a precision, so each
DB defaults to a certain number of digits. This number is 6 in Oracle, and the
value is round up internally instead of being truncated.

Explicitly casting the datetime literal in the WHERE clause causes the appropriate
comparison to occur.

A new unit test specifically for this issue showed that the same issue was also
occurring in certain H2 modes, causing flaky local tests.