fix: EXPOSED-123 ExposedBlob.getBytes() fails on Oracle with IOException Mark invalid
The following tests fail when run using Oracle:

DDLTests/testBlob()
DDLTests/testBlobDefault()
EntityTests/testBlobField()

They all fail with the same error:
Caused by: java.io.IOException: Mark invalid or stream not marked.
	at oracle.jdbc.driver.OracleBlobInputStream.reset(OracleBlobInputStream.java:308)
	at org.jetbrains.exposed.sql.statements.api.ExposedBlob.getBytes(ExposedBlob.kt:13)

The exception is thrown after a BLOB value is retrieved and read from the DB because
the type retrieved is oracle.jdbc.driver.OracleBlobInputStream, which either does
not set markedByte to not be default by calling mark() or invalidates mark. This
makes reset() throw.

This is unlike an ExposedBlob instantiated using the secondary constructor that
creates a ByteArrayInputStream from a ByteArray, which overrides reset() to not throw.

This exception is handled inside getBytes() to ensure that the retrieved BLOB
value is still read as a stream.