fix: EXPOSED-128 Update with Join and Where clause fails in Oracle
The following test fails when run on Oracle:

UpdateTests/testUpdateWithJoinAndWhere()

Fails with an assertion error because the arguments passed to PreparedStatement
are in the wrong positions.

This happens because Oracle is the only supported database to use a subquery directly
in the UPDATE syntax to represent the join part and the WHERE clause together.
And, unlike other DB, this WHERE clause also comes before the UPDATE set clause.

Fixed by reversing the order of arguments in the UpdateStatement query builder,
specifically when an update-with-join operation is executed.