fix: EXPOSED-135 Oracle does not use setSchema value as currentScheme
The following test fails when run on Oracle:

CreateIndexTests/'test possibility to create indexes when table exists in different schemas'()

Fails with 'ORA-01430: column being added already exists in table' due to this
statement being incorrectly generated:
ALTER TABLE TEST_TABLE ADD ID NUMBER(12) NOT NULL

This happens because metadata returns an empty map of existing table ColumnMetadata
when columns() is invoked in JdbcDatabaseMetadataImpl. getColumns() relies on the
value of currentScheme which is invariably set to return databaseName in Oracle;
that is, the default user set up at connection, EXPOSEDTEST.

So, even though a correct statement is executed, which changes schema in the DB,
Exposed is using the default schema to query metadata.

currentScheme now attempts to get a connection schema first before using databaseName.