fix: EXPOSED-122 Fix timestampWithTimeZone tests in Oracle
The following test fails when run on Oracle in the 3 modules:
exposed-kotlin-datetime, exposed-java-time, exposed-jodatime

DefaultsTest/testTimestampWithTimeZoneDefaults

The following tests fail when run on Oracle in exposed-jodatime only:

JodaTimeBaseTest/testTimestampWithTimeZone
JodaTimeMiscTableTest/testTimestampWithTimeZone

Issue 1:
Fails with:
ORA-01861: literal does not match format string
Because Oracle is set to use the default formatter (2023-08-09T11:26:32.7367231Z)
instead of the expected pattern (2023-08-09 11:26:32.7367231 +00:00).

Oracle now has its own datetime formatter in each module.

Issue 2:

In exposed-jodatime module only, reading a timestamp with time zone object in Oracle
then fails with a NPE. This occurs because of a read conversion error since DateTime
is sent to the DB as java.sql.Timestamp and ends up being stored as oracle.sql.TIMESTAMPTZ.

The object is instead read and retrieved as a java.sql.Timestamp.