fix: EXPOSED-137 SET DEFAULT option should not be supported in Oracle
The following test fails when run on Oracle:

sqlite/ForeignKeyConstraintTests/'test ON DELETE SET DEFAULT for databases that support it without SQLite'()

Fails with:
java.sql.SQLException: ORA-03001: unimplemented feature

Oracle's ON DELETE clause only supports CASCADE or SET NULL reference options.

A warning is now logged and the option is not appended to the query string.