test: Fix failing Oracle exposed-core tests
The following tests fail when run on Oracle:

CreateMissingTablesAndColumnsTests/testCamelCaseForeignKeyCreation()

Fails with:
ORA-02275: such a referential constraint already exists in the table.

Because Oracle metadata (via getImportedKeys()) only returns foreign keys that
reference primary keys of the referenced table.

The test now excludes Oracle.

EntityTests/preloadOptionalReferencesOnASizedIterable()

Fails on the first assertNotNull() because testCache() is unable to find
school1.id = 0, which does not exist. The entity school1 is cached with its correct
id.value = 1, but an incorrect value is passed as an argument unless the entity
is first flushed after creation.