fix: EXPOSED-130 Logger throws ClassCastException with JSON and ListSerializer
When a json/jsonb column is created that takes a List or ListSerializer, it correctly
deserializes/serializes the json array to and from the database. However, attempting
to log the statements involved throws ClassCastException, whereas there is no
issue if the column is created to take an Array or ArraySerializer instead.

The former extend Iterable, which is being handled separately in ColumnType.valueToString().
Overriding this branch to directly use the result of nonNullValueToString() fixes
the logger issue.