fix: EXPOSED-133 Suspend transactions blocking Hikari connection pool
If the maximumPoolSize of a HikariDataSource is reached, the active connections
made through newSuspendedTransaction() are not being released back into the pool,
leading to a connection timeout exception when attempting to get more connections.

resetIfClosed(), in suspendedTransactionAsyncInternal(), calls getConnection() to
check if the transaction is closed (may happen after a repetition attempt), so
that it can be properly reset for a new attempt.

This blocks the connections being released and should not be necessary for the
first loop, as the created transaction has an open connection.

resetIfClosed() is now only accessed on subsequent repetition attempt loops.