fix: Verbose logging introduced by LogCaptor
Adding the LogCaptor dependency caused the logging level to be DEBUG instead of INFO, even though it is configured to be INFO in log4j2.xml. This is a known issue (https://github.com/Hakky54/log-captor/issues/8) because LogCaptor uses a different appender than the one configured in log4j2.xml. More details in this comment: https://github.com/Hakky54/log-captor/issues/8#issuecomment-850676609.

The solution is to remove the dependency and create a custom appender which is used to test the logs.