fix: EXPOSED-178 DELETE_RULE read incorrectly for Oracle
Oracle returns CASCADE=0, SET_NULL=2, and NO_ACTION=1
`decode (f.delete_rule, 'CASCADE', 0, 'SET NULL', 2, 1) as delete_rule`

So when it returns 1, which in the current code corresponds to DatabaseMetaData.importedKeyRestrict, it will now be mapped to NO_ACTION because Oracle does not have the RESTRICT reference option. NO_ACTION behaves likes RESTRICT for it.