deprecate: Raise deprecation level of currentScheme property
Raise deprecation level of ExposedDatabaseMetadata.currentScheme from ERROR to
WARNING.

This property was meant to be private/protected within classes that implement
the interface, but has been used in VendorDialect` and 1 unit test.

A new abstract fun tableNamesByCurrentSchema() is introduced to abstract away
the schema from the core module. Ideally this function would return a List<String>
but tableExists() needs to know about the current scheme name, so it instead
returns a Pair to provide that.