deprecate: Raise deprecation level of currentScheme property (#1874)
* deprecate: Raise deprecation level of currentScheme property

Raise deprecation level of ExposedDatabaseMetadata.currentScheme from WARNING to
ERROR.

This property was initially deprecated after PR 815 introduced the possibility to change the 
schema in runtime.

The goal of this fix is to keep the property private in JdbcDatabaseMetadataImpl, but it has 
been used in VendorDialect` and 1 unit test.

A new abstract fun tableNamesByCurrentSchema() is introduced to abstract away
the schema from the core module. Ideally this function would return a List<String>
but tableExists() needs to know about the current scheme name. So, instead, a new data 
class SchemaMetadata is introduced to store both the schema name and its list of table names.