refactor: EXPOSED-65 Change function declaration of adjustSelect
The new query DSL only changed the name of the function from adjustSlice()
to adjustSelect(). Since the new select() returns a Query instead of a FieldSet,
users would need to access the set property within the lambda expression to
resolve type mismatch errors.

Changing the function declaration to accept a lambda block that returns a Query
avoids placing this responsibility on the users during migration, since the
recommendation had always been to use this function with a slice().

Users that have been providing a FieldSet like a Table in the lambda block
will need to add selectAll() to create a query.