perf: EXPOSED-204 Performance problem with getConnection()
In spite of reused connections from a connection pool instance, the lazy
property initializer of a new database connection attempts to reset the
transaction isolation level. This occurs regardless of whether the connection
pool's set level has been overriden by a new value setting in DatabaseConfig
or on the transaction itself.

A flag in Database is now set if a DataSource is used to setup connections.
On the first connection, this flag ensures that the level set by the DataSource
is retrieved and cached. This avoids redundant resets in future connections
unless the DataSource value has been overriden by DatabaseConfig or per-transaction
values.