fix: EXPOSED-242 [PostgreSQL] Cannot change setting in middle of a transaction
Exposed attempts to reset connection settings (isolation level and read only) every
time a new transaction is created, if the transaction settings don't match the
default settings. This is regardless of whether a connection pool is being used
with its own connection settings.

The first half of this issue has already been fixed (PR 1943) so the next exception
thrown becomes: "Cannot change transaction read only in the middle of a transaction."

The same cache and check implementation for that fix has now been applied to the
read only setting when a new transaction is created.