docs: Edit KDocs for Op.TRUE/FALSE to clarify purpose (#1957)
Current KDocs are potentially misleading in that it seems like the operator
could be used both as a standalone value and on the right-hand side of a
comparison operator. For most databases, this works, but will lead to a syntax
error in SQL Server and Oracle.

The new KDocs clarify that they represent operators that always evaluate to a
boolean value, so using them on the right-hand side is not the intention.