fix: EXPOSED-257 Upsert incorectly parameterizes non-literal WHERE arguments
Currently when using upsert() with a where argument, only literal values generate
valid SQL. Attempting to use a value with a param() wrapper or the plain native
type results in an exception that the parameter count is incorrect.

This occurs because the where arguments were not correctly added to the list of
statement arguments in the super class, so an override has now been included in
UpsertStatement directly.