fix: EXPOSED-272 [MySQL, Oracle] Unsupported type BIGINT UNSIGNED for auto-increment
With both Oracle and MySQL, attempting to create a table with an unsigned integer
or long column that should auto-increment throws an `Unsupported type` exception.

This occurs because `AutoIncColumnType`, which delegates to the underlying type,
is currently set up to only detect signed integer and long column types.

It is not possible to reuse the existing `integerAutoincType()` or `longAutoincType()`
if the delegate is an unsigned column because of the type syntax differences
between standard and auto-incrementing columns in some database.
So new types have been added to `DataTypeProvider`.