fix: EXPOSED-266 Between() accepts arguments of different type than column type
between() currently accepts arguments of types that do not match the type stored
in the column on which it is called. In the event that this is being taken advantage
of by users, rather than introducing an immediate breaking change, the upper bounds
have been restricted to show a compiler warning.

A function override has also been introduced to ensure that this warning does not
show when invoked on an EntityID column with literal arguments.