fix: EXPOSED-287 Wrong parenthesis with advanced use of isDistinctFrom
Attempting to use either operator isDistinctFrom() or isNotDistinctFrom() with a
left-or right-side value that is a complex expression results in either a syntax
exception or incorrect results (depending on the database).

Complex expressions like isNull() and isNotNull() are already marked, so this fix
ensures that they are identified as such and wrapped in parentheses when appending
to a query builder, by using appendExpression() instead of base append().