fix: EXPOSED-292 Explicit nulls in insert with databaseGenerated()
If a generated/computed column is marked as such using databaseGenerated(), but it is also nullable(), an exception is thrown even when the field is correctly excluded in an insert() block.

This occurs because a NULL value is still being auto-included for that column in the prepared insert statement.

This fix ensures that the column is excluded from adding an SQL NULL if it is marked
as database generated. The class property is also included wherever a column is copied
to ensure that swapping chained calls produces the same result.