feat: EXPOSED-258 Enhance upsert to allow exclusion of columns set on conflict
Currently, if the onUpdate parameter of upsert() is provided an argument, these
column-value pairs will be used in the update clause of the upsert statement.
If the argument is left null, a duplicate of the column-value assignments from
the insert block will be used instead.
If a user wants to only update a subset of columns, or not update columns with
default values, the only option is to provide an argument to onUpdate that
duplicates most of the insert block.

This can become verbose, so an additional parameter, onUpdateExclude, has been
introduced, which duplicates all insert column-value pairs except those for the
specifically excluded columns.
This parameter could also be used in the case when the number of columns to update
is small but the user wants to avoid duplicating the insert values.