fix: EXPOSED-301 Update with join throws if additionalConstraint provided
Attempting to use any join function in its full overload form, with an additionalConstraint
argument, results in an exception if an update() is called instead of a select().

This occurs because the additionalConstraint conditions are being correctly appended
to generated SQL, but their arguments are not being first registered with the prepared
update statement.

The arguments in the update statement now correctly register any additionalConstraint
and take into account their order if an update with where is called.