fix: EXPOSED-217 Unnecessary query after calling with() and iteration
When eager loading references from a SizedIterable of entities, with() first iterates
over the object then preloads relations, which correctly generates 2 SQL queries.

If this eager loading is followed by iteration, the first SQL query is executed again
because its results are discarded, while the retrieved references are cached.

This fix stores the result of the first query in the SizedIterable instance to
ensure that the right iterator is invoked.