fix!: EXPOSED-317 repetitionAttempts property is misleading (#2042)
* fix: EXPOSED-317 repetitionAttempts property is misleading

The KDocs for the Transaction property repetitionAttempts makes it seem like
setting the value to 1 would mean that any failed transaction should undergo 1
automatic retry. Instead, setting the value to 1 means that the transaction is
only run once and any failure is thrown.

Given that both the KDocs and the use of repetition in the name is potentially
misleading, the property and its related delay and default variants will be deprecated
in favor of more aligned and conventional properties:
- Add maxAttempts, minRetryDelay, and maxRetryDelay and default variants
- Set existing properties to DeprecationLevel.WARNING
- Update tests
- Update documentation

* fix: EXPOSED-317 repetitionAttempts property is misleading

Add getters and setters to deprecated properties so the new properties are the ssot.
Add checks for attempts not being set a value less than 1.