fix: EXPOSED-340 Syntax error using upsert with MySQL8 below 8.0.19
MySQL `upsert()` uses an alias for the insert row values when the database
version is 8+.
Technically this is only possible starting with version 8.0.19.

Existing function `MysqlDialect.isMysql8` calls `isVersionCovers()`, which only
fetches the database's major and minor version, so it is not sufficient for
this comparison.
To get the patch version as well requires fetching metadata using the full
`databaseProductVersion` property.