deprecate: EXPOSED-354 Database.connectPool() with ConnectionPoolDataSource
Database.connect() has an alternative, Database.connectPool() that takes a
ConnectionPoolDataSource as an argument. JDBC driver implementations of this class
are not actual connection pools, but rather a means of obtaining the physical
connections to be used in a connection pool.

Third party connection pools may provide a better alternative than relying on
the JDBC driver to pool connections.

connectPool() has been set to DeprecationLevel.WARNING.